pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.22",
    "google-genai>=1.64.0",
    "cerebras-cloud-sdk>=1.67.0",
    "psycopg2-binary>=2.9.11",
]
